Working (Login.php and Register.php)
Login works while also setting a session token to stay logged in.

Register works but currently does not implement an email.

Currently the only way to destroy the session is through logout.php. Will implement into a button in the future once the main page is created.

TODO:
Add email into registration
Allow for logging in with either email or username.